Bug 835635: Cast ArrayLength() result to (signed) 'gint' type in #define for CUSTOM_VALUE_INDEX, since we compare it to gint values all over the place. r=karlt

+++ widget/gtk2/nsPrintDialogGTK.cpp
@@ -36,7 +36,7 @@ using namespace mozilla::widget;
 
 static const char header_footer_tags[][4] =  {"", "&T", "&U", "&D", "&P", "&PT"};
 
-#define CUSTOM_VALUE_INDEX ArrayLength(header_footer_tags)
+#define CUSTOM_VALUE_INDEX gint(ArrayLength(header_footer_tags))
 
 static GtkWindow *
 get_gtk_window_for_nsiwidget(nsIWidget *widget)
